+++ content/pages/now.mdx
@@ -1,0 +1,26 @@
+---
+title: "Now Page"
+date: "2021-07-11"
+description: "An overview of what's going on in my life at the moment and what I'm working on "
+---
+
+## Current Status
+
+I am currently seeking employment as a FullStack Developer. I've met a couple of rejections here and there,
+I remain optimistic regardless that soon enough something will come my way. If you would like to work together
+feel free to send me an email at [inezabonte@gmail.com](mailto:inezabonte@gmail.com)
+
+## What I'm currently working on
+
+I've been working on this portfolio for most part of my time. I'm really happy with what I've been able
+to come up with. I have quite a list of issues pending which I shall implement gradually as I continue
+to learn more about Next.js and TailwindCSS.
+I've also been learning TypeScript after realising how important it is in the developement process
+and also the fact that it's an in-demand skill.
+
+## Anything else?
+
+Be kind to one another. ❤️
+
+
+Last Updated: 13th July 2021
