+++ lib/articles.js
@@ -1,9 +1,6 @@
 import fs from "fs";
 import path from "path";
 import matter from "gray-matter";
-import readingTime from "reading-time";
-import { serialize } from "next-mdx-remote/serialize";
-import mdxPrism from "mdx-prism";
 
 const articlesDirectory = path.join(process.cwd(), "content/articles");
 
@@ -41,30 +38,4 @@ export function getAllArticleIds() {
 			},
 		};
 	});
-}
-
-export async function getArticleData(id) {
-	const fullPath = path.join(articlesDirectory, `${id}.mdx`);
-	const fileContents = fs.readFileSync(fullPath, "utf-8");
-
-	const { data, content } = matter(fileContents);
-	const mdxSource = await serialize(content, {
-		mdxOptions: {
-			remarkPlugins: [
-				require("remark-code-titles"),
-				require("remark-autolink-headings"),
-			],
-			rehypePlugins: [mdxPrism],
-		},
-		scope: data,
-	});
-
-	const readTime = readingTime(content);
-
-	return {
-		id,
-		mdxSource,
-		readTime,
-		frontMatter: data,
-	};
 }
