+++ package.json
@@ -24,8 +24,7 @@
     "react-icons": "^4.2.0",
     "reading-time": "^1.3.0",
     "rehype": "^11.0.0",
-    "remark-autolink-headings": "^6.0.1",
-    "remark-code-titles": "^0.1.2"
+    "remark-autolink-headings": "^6.0.1"
   },
   "devDependencies": {
     "autoprefixer": "^10.2.5",
